Fix travis_url variable name bug & improve lookup
The `travis_url` key from `projects.json` was being referred to as `travis url`.

I also added a check for the scenario where the guid has been renamed and it can't find a match between the guid in the DB and the one in `projects.json`. Instead of crashing, the app will just ignore the mismatch.